Bughound
The online website is developed, which provides the user the ease to add
a new bug report, view a bug report and update a bug report. “Bughound”
is currently a standalone website. The user can insert a new bug report
in the database, user can view the bug details and even update the
previously submitted bug report. This application is developed in PHP
with SQLWorkbench as its storage manager.